Support passing user-supplied data to TfLiteOperator's methods
The current mechanism of having TfLiteOperator.init return a user-allocated buffer
doesn't allow for the TfLiteOperator.init to take a user-supplied pointer. And that is something we want to support so as to allow the developer to pass  TfLiteOperatorCreate() a C struct or C++ class implementing the Init/Free/Prepare/Eval methods.

PiperOrigin-RevId: 630560524